Enable using the Soloplan.Build NuGet on the build server
With this change, it's possible to use the Soloplan.Build.NuGet from by setting the MSBuild Property UseSoloBuildNuget=True. By introducing this condition, it's still possible to work with the project locally.
This allows that Soloplan can sign the output binaries for the official releases of this repository.

Set the InternalsVisibleTo assembly property depending on whether the assemblies are signed or not.