Configure connectors with a string identifier
Previously this was bound to the name of a CLR type which bound the configuration directly to the programming model.
Now, it's configured with a unique identifier, defined on the Connector implementation on the particular plugins.

Rename JenkinsProjectPlugin to JenkinsPlugin.

Introduce a hardcoded configuration to stay compatible to old configurations (will be dropped with the 1.0 release).

Drop the IConnectorPlugin interface and use the abstract ConnectorPlugin class instead.