Remove the "Edit (wizard)" button from the configuration.
This concept, combined with the "New (wizard)" feature is confusing to the user. What does edit actually do compared to new? Why do I have to select a project from a server if I want to edit a project? These questions and more arise when looking at this feature.
On top, if you clean out the address field and click the edit button, the wizard will just crash.

The wizard supports the process of adding new projects. Using it for adjusting the project name in the same manner is confusing and miss-usage of its original purpose.
The actual feature of "Use a new project from the wizard to configure my existing project configuration" is not something very useful. If a user wants to "re-add" a project from the wizard, he should just remove it and use the "Add project" function.

Rename "New (wizard)" to "Add project" to comply withe the wording for opening the wizard from the main window.